requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
hydra-core>=1.3.0
hydra-colorlog>=1.2.0
wandb>=0.15.0
omegaconf>=2.3.0
datasets>=2.14.0
transformers>=4.30.0
tokenizers>=0.13.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0
tensorboard>=2.13.0
pillow>=10.0.0
scikit-learn>=1.3.0
accelerate>=0.20.0
peft>=0.4.0
bitsandbytes>=0.41.0 